% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tokens.R
\name{count_tokens}
\alias{count_tokens}
\title{Find frequently occurring tokens within a hierarchical column}
\usage{
count_tokens(
  x,
  split = "[-_[:space:]]+",
  min_freq = 2,
  min_nchar = 3,
  return_values = TRUE,
  std_fn = string_std,
  ...
)
}
\arguments{
\item{x}{a character vector (generally a hierarchical column)}

\item{split}{regex pattern used to split values into tokens. By default
splits on any sequence of one or more space characters ("[:space:]"),
dashes ("-"), and/or underscores ("_").}

\item{min_freq}{minimum token frequency (i.e. number of unique values in
which a given token occurs). Defaults to \code{2}.}

\item{min_nchar}{minimum token size in number of characters. Defaults to \code{3}.}

\item{return_values}{logical indicating whether to return the standardized
values in which each token is found (\code{TRUE}), or only the count of the
number of unique standardized values (\code{FALSE}). Defaults to \code{TRUE}.}

\item{std_fn}{function to standardize strings, as performed within all
\code{hmatch_} functions. Defaults to \code{\link{string_std}}. Set to \code{NULL} to
omit standardization. See also \link{string_standardization}.}

\item{...}{additional arguments passed to \code{std_fn()}}
}
\description{
Tokenized matching of hierarchical columns can yield false positives when
there are tokens that occur frequently in multiple unique hierarchical values
(e.g. "South", "North", "City", etc.).

This is a helper function to find such frequently-occurring tokens, which can
then be passed to the \code{exclude} argument of \code{\link{hmatch_tokens}}. The
frequency calculated is the number of unique,
\link[=string_standardization]{string-standardized} values in which a given
token is found.
}
\examples{
french_departments <- c(
  "Alpes-de-Haute-Provence", "Hautes-Alpes", "Ardennes", "Bouches-du-Rhône",
  "Corse-du-Sud", "Haute-Corse", "Haute-Garonne", "Ille-et-Vilaine",
  "Haute-Loire", "Hautes-Pyrénées", "Pyrénées-Atlantiques", "Hauts-de-Seine"
)

count_tokens(french_departments)

}
