% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drc_man}
\alias{drc_man}
\title{Dataset of manual corrections}
\format{A data.frame with 1 row and 5 variables:
\describe{
\item{adm1}{Name of administrative 1 level}
\item{adm2}{Name of administrative 2 level}
\item{adm3}{Name of administrative 3 level}
\item{adm4}{Name of administrative 4 level}
\item{pcode}{Geocode}
}}
\usage{
drc_man
}
\description{
Raw entries of adminstrative districts from the Democratic Republic of the
Congo, and their corresponding geocode.
}
\keyword{datasets}
