% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drc_raw}
\alias{drc_raw}
\title{Raw dataset}
\format{A data.frame with 11 rows and 5 variables:
\describe{
\item{id}{Integer id of entry}
\item{adm1}{Name of administrative 1 level}
\item{adm2}{Name of administrative 2 level}
\item{adm3}{Name of administrative 3 level}
\item{adm4}{Name of administrative 4 level}
}}
\usage{
drc_raw
}
\description{
Raw entries of adminstrative districts from the Democratic Republic of the
Congo.
}
\keyword{datasets}
