% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drc_ref}
\alias{drc_ref}
\title{Reference dataset}
\format{A data.frame with 15 rows and 6 variables:
\describe{
\item{level}{Level of geocode}
\item{adm1}{Name of administrative 1 level}
\item{adm2}{Name of administrative 2 level}
\item{adm3}{Name of administrative 3 level}
\item{adm4}{Name of administrative 4 level}
\item{pcode}{Geocode}
}}
\usage{
drc_ref
}
\description{
Reference table of select adminstrative districts from the Democratic
Republic of the Congo.
}
\keyword{datasets}
