% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcodes.R
\name{hcodes}
\alias{hcodes}
\alias{hcodes_str}
\alias{hcodes_int}
\title{Create codes to identify each unique combination of hierarchical levels in
a reference dataset}
\usage{
hcodes_str(ref, pattern, by, sep = "__", std_fn = string_std)

hcodes_int(ref, pattern, by, prefix = "")
}
\arguments{
\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern}{regex pattern to match the names of the hierarchical columns
in \code{ref} (supply either \code{pattern} \emph{or} \code{by})}

\item{by}{vector giving the names of the hierarchical columns in \code{ref}
(supply either \code{pattern} \emph{or} \code{by})}

\item{sep}{(only for \code{hcodes_str}) desired separator between levels in
string-based codes (defaults to "__")}

\item{std_fn}{(only for \code{hcodes_str}) Function to standardize input strings
prior to creating codes. Defaults to \code{\link{string_std}}. Set to
\code{NULL} to omit standardization. See also \link{string_standardization}.}

\item{prefix}{(only for \code{hcodes_int}) character prefix for integer-based
codes (defaults to "")}
}
\value{
A vector of codes
}
\description{
Create codes to identify each unique combination of hierarchical levels in a
reference dataset. Codes may be integer-based (function \code{hcodes_int}) or
string-based (\code{hcodes_str}). Integer-based codes reflect the alphabetical
ranking of each level within the next-highest level. They are constant-width
and may optionally be prefixed with any given string. String-based codes are
created by pasting together the values of each hierarchical level with a
given separator (with options for string standardization prior to
collapsing).
}
\examples{
data(ne_ref)

# string-based codes
hcodes_str(ne_ref, pattern = "^adm")

# integer-based codes
hcodes_int(ne_ref, pattern = "^adm")

}
