% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch.R
\name{hmatch}
\alias{hmatch}
\title{Implement a variety of hierarchical matching strategies in sequence}
\usage{
hmatch(
  raw,
  ref,
  man = NULL,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  by = NULL,
  code_col = NULL,
  std_fn = string_std,
  fuzzy = FALSE,
  max_dist = 1L
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{man}{(optional) \code{data.frame} of manually-specified matches, relating a
given set of hierarchical values to the code within \code{ref} to which those
values correspond}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}
(and \code{man} if given) (see also \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(see also \link{specifying_columns})}

\item{by}{named character vector whose elements are the names of the
hierarchical columns in \code{ref}, and whose names are the names of the
corresponding columns in \code{raw} (and \code{man}, if given) (see also
\link{specifying_columns})}

\item{code_col}{name of the code column containing codes for matching \code{ref}
and \code{man} (only required if argument \code{man} is given)}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}
}
\value{
A \code{data.frame} obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, based on the matches specified in \code{man}. If \code{type == "inner"},
returns only the rows of \code{raw} with a single match in \code{ref}. If \code{type == "left"}, returns all rows of \code{raw}. If the hierarchical columns within
\code{ref} have identical names to \code{raw}, the returned reference columns will be
renamed with prefix "bind_".
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using a variety of
matching strategies implemented in sequence to identify the best-possible
match (i.e. highest-resolution) for each row.

The sequence of matching stragies is:
\enumerate{
\item (optional) manually-specified matching with \code{\link{hmatch_manual}}
\item complete matching with \code{\link{hmatch_complete}}
\item partial matching with \code{\link{hmatch_partial}}
\item fuzzy partial matching with \code{\link{hmatch_partial}}
\item best-possible matching with \code{\link{hmatch_best}}
}

Each approach is implement only on the rows of data for which a single match
has not already been identified using the previous approaches.
}
\examples{
data(ne_raw)
data(ne_ref)

hmatch(ne_raw, ne_ref, fuzzy = TRUE)

}
