% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_composite.R
\name{hmatch_composite}
\alias{hmatch_composite}
\title{Implement a variety of hierarchical matching strategies in sequence}
\usage{
hmatch_composite(
  raw,
  ref,
  man = NULL,
  pattern = NULL,
  pattern_ref = pattern,
  by = NULL,
  by_ref = by,
  code_col = NULL,
  type = "resolve_left",
  allow_gaps = TRUE,
  fuzzy = FALSE,
  max_dist = 1L,
  dict = NULL,
  ref_prefix = "ref_",
  std_fn = string_std,
  ...
)
}
\arguments{
\item{raw}{data frame containing hierarchical columns with raw data}

\item{ref}{data frame containing hierarchical columns with reference data}

\item{man}{(optional) data frame of manually-specified matches, relating a
given set of hierarchical values to the code within \code{ref} to which those
values correspond}

\item{pattern}{regex pattern to match the hierarchical columns in \code{raw}
(and \code{man} if given) (see also \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(defaults to \code{pattern})}

\item{by}{vector giving the names of the hierarchical columns in \code{raw} (and
\code{man} if given)}

\item{by_ref}{vector giving the names of the hierarchical columns in \code{ref}
(defaults to \code{by})}

\item{code_col}{name of the code column containing codes for matching \code{ref}
and \code{man} (only required if argument \code{man} is given)}

\item{type}{type of join ("resolve_left", "resolve_inner", or
"resolve_anti"). Defaults to "left". See \link{join_types}.}

\item{allow_gaps}{logical indicating whether to allow missing values below
the match level (defaults to \code{TRUE})}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}

\item{dict}{optional dictionary for recoding values within the hierarchical
columns of \code{raw} (see \link{dictionary_recoding})}

\item{ref_prefix}{prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to "ref_")}

\item{std_fn}{function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{additional arguments passed to \code{std_fn()}}
}
\value{
a data frame obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, using the join type specified by argument \code{type} (see
\link{join_types} for more details)
}
\description{
Match a data frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using a variety of
matching strategies implemented in sequence to identify the best-possible
match (i.e. highest-resolution) for each row.

The sequence of matching strategies is:
\enumerate{
\item (optional) manually-specified matching with \code{\link{hmatch_manual}}
\item complete matching with \code{\link{hmatch_partial}} (\code{allow_gaps = FALSE})
\item partial matching with \code{\link{hmatch_partial}} (\code{allow_gaps = TRUE})
\item fuzzy partial matching with \code{\link{hmatch_partial}} (\code{allow_gaps = TRUE}, \code{fuzzy = TRUE})
\item best-possible matching with \code{\link{hmatch_best}}
}

Each approach is implement only on the rows of data for which a single match
has not already been identified using the previous approaches.
}
\examples{
data(ne_raw)
data(ne_ref)

hmatch_composite(ne_raw, ne_ref, fuzzy = TRUE)

# with dictionary-based recoding
ne_dict <- data.frame(value = "USA",
                      replacement = "United States",
                      variable = "adm0")

hmatch_composite(ne_raw, ne_ref, dict = ne_dict, fuzzy = TRUE)

}
