% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_exact.R
\name{hmatch_exact}
\alias{hmatch_exact}
\title{Exact hierarchical matching}
\usage{
hmatch_exact(
  raw,
  ref,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  by = NULL,
  type = "left",
  std_fn = string_std
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}}

\item{by}{named character vector whose elements are the names of the
hierarchical columns in \code{ref} and whose names are the names of the
corresponding columns in \code{raw}}

\item{type}{type of join ("left", "inner", "anti") (defaults to "left")}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}
}
\value{
A \code{data.frame} obtained by matching the hierarchical columns in \code{raw}
and \code{ref}. If \code{type == "inner"}, returns only the rows of \code{raw} with a
single match in \code{ref}. If \code{type == "left"}, returns all rows of \code{raw}. If
the hierarchical columns within \code{ref} have identical names to \code{raw}, the
returned reference columns will be renamed with prefix "bind_".
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using exact
matching. "Exact" here means that every level must match in sequence. For
example, given raw data with known levels of county and township, but missing
the lower-resolution level province, a match will not be made.
}
\examples{
data(ne_raw)
data(ne_ref)

hmatch_exact(ne_raw, ne_ref, pattern_raw = "adm", type = "inner")

}
