% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_manual.R
\name{hmatch_manual}
\alias{hmatch_manual}
\title{Manual hierarchical matching}
\usage{
hmatch_manual(
  raw,
  ref,
  man,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  pattern_man = pattern_raw,
  by = NULL,
  code_col = "pcode",
  type = "left"
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{man}{\code{data.frame} of manually-specified matches, relating a given set
of hierarchical values to the code within \code{ref} to which those values
correspond}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}}

\item{pattern_man}{regex pattern to match the hierarchical columns in \code{man}}

\item{by}{named character vector whose elements are the names of the
hierarchical columns in \code{raw} and \code{man} and whose names are the names of
the corresponding columns in \code{ref}}

\item{code_col}{name of the code column within \code{ref} and \code{man}}

\item{type}{type of join ("inner" or "left") (defaults to "left")}
}
\value{
A \code{data.frame} obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, based on the matches specified in \code{man}. If \code{type == "inner"},
returns only the rows of \code{raw} with a single match in \code{ref}. If \code{type == "left"}, returns all rows of \code{raw}. If the hierarchical columns within
\code{ref} have identical names to \code{raw}, the output reference columns will be
renamed with prefix "bind_".
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using a dictionary
of manually-specified matches.
}
\examples{
data(drc_raw)
data(drc_ref)
data(drc_man)

hmatch_manual(drc_raw, drc_ref, drc_man, code_col = "pcode")

}
