% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_levels.R
\name{max_levels}
\alias{max_levels}
\title{Maximum hierarchical levels}
\usage{
max_levels(x, pattern, by, type = c("index", "name"))
}
\arguments{
\item{x}{a data frame containing hierarchical columns}

\item{pattern}{regex pattern to match the names of the hierarchical columns
in \code{ref} (supply either \code{pattern} \emph{or} \code{by})}

\item{by}{vector giving the names of the hierarchical columns in \code{ref}
(supply either \code{pattern} \emph{or} \code{by})}

\item{type}{type of return, either "index" to return integer indices
(starting at 1) or "name" to return column names (as matched by \code{pattern}
or \code{by})}
}
\value{
Vector of indices or names corresponding to the maximum non-missing
hierarchical level for each row
}
\description{
Given a data frame with columns specifying hierarchically-nested levels, find
the maximum non-missing hierarchical level for each row.
}
\examples{
data(ne_ref)

# return integer indices (starting at 1)
max_levels(ne_raw, pattern = "^adm")

# return column names
max_levels(ne_raw, pattern = "^adm", type = "name")

}
