% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ne_ref}
\alias{ne_ref}
\title{Reference dataset}
\format{
A data.frame with 31 rows and 4 variables, all of class character:
\describe{
\item{level}{Administrative level}
\item{adm0}{Name of administrative 0 level (country)}
\item{adm1}{Name of administrative 1 level (state/province)}
\item{adm2}{Name of administrative 2 level (county/census division)}
\item{hcode}{Hierarchical code}
}
}
\usage{
ne_ref
}
\description{
Reference table of select administrative districts in the northeastern
portion of North America.
}
\keyword{datasets}
