% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_hcode.R
\name{separate_hcode}
\alias{separate_hcode}
\title{Separate a hierarchical code reflecting multiple levels into its constituent
parts, with one column for each level}
\usage{
separate_hcode(
  x,
  ref = NULL,
  by = NULL,
  col,
  into,
  sep = "__",
  extra = c("warn", "drop"),
  remove = FALSE
)
}
\arguments{
\item{x}{\code{data.frame} containing a column with hierarchical codes}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{by}{vector giving the names of the hierarchical columns in \code{ref}}

\item{col}{Name of the column within \code{x} containing hierarchical codes.}

\item{into}{Vector of column names to separate \code{col} into}

\item{sep}{Separator between levels in the hierarchical codes. Defaults to
"__".}

\item{extra}{What to do if a hierarchical code contains more levels than are
implied by argument \code{into}.
\itemize{
\item "warn" (the default): emit a warning and drop extra values
\item "drop": drop any extra values without a warning
}}

\item{remove}{Logical indicating whether to remove \code{col} from the output.
Defaults to \code{FALSE}.}
}
\value{
The original data.frame \code{x} with additional columns for each level of the
hierarchical code
}
\description{
Separate a data frame column containing hierarchical codes into multiple
columns, one for each level within the hierarchical code.

Like \code{\link[tidyr:separate]{tidyr::separate}} except that successive levels are cumulative rather
then independent. E.g. the code "canada__ontario__toronto" would be split
into three levels:
\enumerate{
\item "canada"
\item "canada__ontario"
\item "canada__ontario__toronto"
}
}
\examples{
data(ne_ref)

# generate pcode
ne_ref$pcode <- hcodes_str(ne_ref, pattern = "^adm\\\\d")

# separate pcode into constituent levels
separate_hcode(
  ne_ref,
  col = "pcode",
  into = c("adm0_pcode", "adm1_pcode", "adm2_pcode")
)

}
