% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmatch_complete.R
\name{spmatch_complete}
\alias{spmatch_complete}
\title{Complete hierarchical matching, separately at each hierarchical level}
\usage{
spmatch_complete(
  raw,
  ref,
  pattern = NULL,
  pattern_ref = pattern,
  by = NULL,
  by_ref = by_ref,
  dict = NULL,
  type = "left",
  ref_prefix = "ref_",
  std_fn = string_std,
  ...,
  levels = NULL,
  always_list = FALSE
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern}{regex pattern to match the hierarchical columns in \code{raw}
(columns can be matched using either the \code{pattern} or \code{by} arguments, as
described in \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(defaults to \code{pattern})}

\item{by}{vector giving the names of the hierarchical columns in \code{raw}}

\item{by_ref}{vector giving the names of the hierarchical columns in \code{ref}
(defaults to \code{by})}

\item{dict}{Optional dictionary for recoding values within the hierarchical
columns of \code{raw} (see \link{dictionary_recoding})}

\item{type}{type of join ("left", "inner", "inner_unique", "anti", or
"anti_unique"). Defaults to "left". See \link{join_types}.}

\item{ref_prefix}{Prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to \code{ref_})}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{Additional arguments passed to \code{std_fn()}}

\item{levels}{a vector of names or integer indices corresponding to one or
more of the hierarchical columns in \code{raw} to match at. Defaults to \code{NULL}
in which case matches are made at each hierarchical level.}

\item{always_list}{logical indicating whether to always return a list, even
when argument \code{levels} specifies a single match level (defaults to \code{FALSE})}
}
\value{
A list of data frames, each returned by a call to \code{hmatch_complete} on the
unique combination of hierarchical values at the given hierarchical level.
The number of elements in the list corresponds to the number of hierarchical
columns in \code{raw}, or, if specified, the number of elements in argument
\code{levels}.

However, if \code{always_list = FALSE} and \code{length(levels) == 1}, a single data
frame is returned (i.e. not wrapped in a list).
}
\description{
Implements hierarchical matching with \code{\link{hmatch_complete}},
separately at each hierarchical level within the data. For a given level, the
raw data that is matched includes every unique combination of values at and
below the level of interest. E.g.

L1: Canada \cr
L1: United States \cr

L2: Canada | Ontario \cr
L2: United States | New York \cr
L2: United States | Pennsylvania \cr

L3: Canada | Ontario | Ottawa \cr
L3: Canada | Ontario | Toronto \cr
L3: United States | New York | New York \cr
L3: United States | Pennsylvania | Philadelphia
}
\examples{
data(ne_raw)
data(ne_ref)

# find all non-matches ("anti"-join) at each hierarchical level
spmatch_complete(ne_raw, ne_ref, type = "anti")

# find all matches ("inner"-join) at only the adm0 level
spmatch_complete(ne_raw, ne_ref, type = "inner", levels = "adm0")

# with dictionary-based recoding
ne_dict <- data.frame(value = "USA",
                      replacement = "United States",
                      variable = "adm0")

spmatch_complete(ne_raw, ne_ref, type = "inner", dict = ne_dict, levels = "adm0")

}
