% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_string_standardization.R
\name{string_standardization}
\alias{string_standardization}
\title{String Standardization}
\description{
Prior to matching raw and reference datasets, one might wish to standardize
the strings within the match columns to account for differences in case,
punctuation, etc.

By default, this standardization is performed with function
\code{\link{string_std}}, which implements four transformations:
\enumerate{
\item standardize case (\code{base::tolower})
\item remove sequences of non-alphanumeric characters at start or end of string
\item replace remaining sequences of non-alphanumeric characters with "_"
\item remove diacritics (\code{stringi::stri_trans_general})
}

Alternatively, the user may provide any function that takes a vector of
strings and returns a vector of transformed strings. To omit any
transformation, set argument \code{std_fn = NULL}.

Note that the standardized versions of the match columns are never returned.
They are used only during matching, and then removed prior to the return.
}
