% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_best.R
\name{hmatch_best}
\alias{hmatch_best}
\title{Find the highest-resolution hierarchical match using a variety of matching
techniques}
\usage{
hmatch_best(
  raw,
  ref,
  man = NULL,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  pattern_man = pattern_raw,
  by = NULL,
  id_col,
  code_col,
  fuzzy = FALSE,
  max_dist = 1L
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{man}{\code{data.frame} of manually-specified matches, relating a given set
of hierarchical values to the code within \code{ref} to which those values
correspond}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}}

\item{pattern_man}{regex pattern to match the hierarchical columns in \code{man}}

\item{by}{named character vector whose elements are the names of the
geo-columns in \code{raw} and whose names are the names of the corresponding
geo-columns in \code{ref} (and \code{man}, if given)}

\item{id_col}{name of the ID column within \code{raw}}

\item{code_col}{name of the code column within \code{ref} (and \code{man}, if given)}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}
}
\value{
A \code{data.frame} obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, based on the matches specified in \code{man}. If \code{type == "inner"},
returns only the rows of \code{raw} with a single match in \code{ref}. If \code{type == "left"}, returns all rows of \code{raw}. If the hierarchical columns within
\code{ref} have identical names to \code{raw}, the output reference columns will be
renamed with prefix "bind_".
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using a
variety of matching strategies to identify the best-possible match (i.e.
highest-resolution) for each row.
}
\examples{
data(drc_raw)
data(drc_ref)

hmatch_best(drc_raw, drc_ref, id_col = "id", code_col = "pcode")

}
