% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_best.R
\name{hmatch_best}
\alias{hmatch_best}
\title{Best-possible hierarchical matching}
\usage{
hmatch_best(
  raw,
  ref,
  pattern = NULL,
  pattern_ref = pattern,
  by = NULL,
  by_ref = by,
  type = "left",
  dict = NULL,
  ref_prefix = "ref_",
  fuzzy = FALSE,
  max_dist = 1L,
  concise = FALSE,
  std_fn = string_std,
  ...
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern}{regex pattern to match the hierarchical columns in \code{raw}
(columns can be matched using either the \code{pattern} or \code{by} arguments, as
described in \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(defaults to \code{pattern})}

\item{by}{vector giving the names of the hierarchical columns in \code{raw}}

\item{by_ref}{vector giving the names of the hierarchical columns in \code{ref}
(defaults to \code{by})}

\item{type}{type of join ("left", "inner", "anti", "inner_complete",
"inner_incomplete"). Defaults to "left". See \link{join_types}.}

\item{dict}{Optional dictionary for recoding values within the hierarchical
columns of \code{raw} (see \link{dictionary_recoding})}

\item{ref_prefix}{Prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to \code{ref_})}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}

\item{concise}{logical indicating whether to return only the hierarchical
columns (from both \code{raw} and \code{ref}) (defaults to \code{FALSE})}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{Additional arguments passed to \code{std_fn()}}
}
\value{
a data frame obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, using the join type specified by argument \code{type} (see
\link{join_types} for more details)

Column \code{match_type} indicates the type of match made:
\itemize{
\item NA: no match
\item "best": the level of the best match from \code{ref} matches the level of \code{raw}
\item "best_low": the level of the best match from \code{ref} is lower than the level
of \code{raw}
\item "best_infer": multiple rows of \code{ref} match \code{raw}; the best match
corresponds to the highest level common among all matches
}
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using a rolling
approach to identify the best-possible (i.e. highest-resolution) match for
each row.

The rolling approach implements \code{\link{hmatch_partial}} at each
successive level, starting with only the first level, then the first and
second level, etc. The 'best-possible' match reflects the highest-level that
is consistent among all possible matches to the given row of raw data.
}
\examples{
data(ne_raw)
data(ne_ref)

hmatch_best(ne_raw, ne_ref, pattern = "adm", fuzzy = TRUE)

# with dictionary-based recoding
ne_dict <- data.frame(value = "USA",
                      replacement = "United States",
                      variable = "adm0")

hmatch_best(ne_raw, ne_ref, dict = ne_dict, fuzzy = TRUE)

}
