% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_complete.R
\name{hmatch_complete}
\alias{hmatch_complete}
\title{Complete hierarchical matching}
\usage{
hmatch_complete(
  raw,
  ref,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  by = NULL,
  dict = NULL,
  type = "left",
  ref_prefix = "ref_",
  std_fn = string_std,
  ...
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}}

\item{by}{named character vector whose elements are the names of the
hierarchical columns in \code{ref} and whose names are the names of the
corresponding columns in \code{raw}}

\item{dict}{Optional dictionary for recoding values within the hierarchical
columns of \code{raw} (see \link{dictionary_recoding})}

\item{type}{type of join ("left", "inner", or "anti"). Defaults to "left".
See \link{join_types}.}

\item{ref_prefix}{Prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to \code{ref_})}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{Additional arguments passed to \code{std_fn()}}
}
\value{
A \code{data.frame} obtained by matching the hierarchical columns in \code{raw}
and \code{ref}. If \code{type == "inner"}, returns only the rows of \code{raw} with a
single match in \code{ref}. If \code{type == "left"}, returns all rows of \code{raw}.
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using complete
matching. "Complete" here means that every level must match in sequence. For
example, given raw data with known levels of county and township, but missing
the lower-resolution level province, a match will not be made.
}
\examples{
data(ne_raw)
data(ne_ref)

hmatch_complete(ne_raw, ne_ref, pattern_raw = "adm")

# with dictionary-based recoding
ne_dict <- data.frame(value = "USA",
                      replacement = "United States",
                      variable = "adm0")

hmatch_complete(ne_raw, ne_ref, pattern_raw = "adm", dict = ne_dict)

}
