% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_manual.R
\name{hmatch_manual}
\alias{hmatch_manual}
\title{Manual hierarchical matching}
\usage{
hmatch_manual(
  raw,
  ref,
  man,
  pattern,
  pattern_ref = pattern,
  by,
  by_ref = by,
  code_col,
  type = "left",
  ref_prefix = "ref_",
  std_fn = string_std,
  ...
)
}
\arguments{
\item{raw}{data frame containing hierarchical columns with raw data}

\item{ref}{data frame containing hierarchical columns with reference data}

\item{man}{\code{data.frame} of manually-specified matches, relating a given set
of hierarchical values to the code within \code{ref} to which those values
correspond}

\item{pattern}{regex pattern to match the hierarchical columns in \code{raw}
and \code{man} (see also \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}.
Defaults to \code{pattern}, so only need to specify if the hierarchical columns
have different names in \code{raw} and \code{ref}.}

\item{by}{vector giving the names of the hierarchical columns in \code{raw} and
\code{man}}

\item{by_ref}{vector giving the names of the hierarchical columns in \code{ref}.
Defaults to \code{by}, so only need to specify if the hierarchical columns
have different names in \code{raw} and \code{ref}.}

\item{code_col}{name of the code column containing codes for matching \code{ref}
and \code{man}}

\item{type}{type of join ("left", "inner", or "anti"). Defaults to "left".
See \link{join_types}. Note that this function does not allow 'resolve
joins', unlike most other \code{hmatch_} functions.}

\item{ref_prefix}{prefix to add to names of returned columns from \code{ref} if
they are otherwise identical to names within \code{raw}. Defaults to "ref_".}

\item{std_fn}{function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{additional arguments passed to \code{std_fn()}}
}
\value{
a data frame obtained by matching the hierarchical columns in \code{raw}
and \code{ref} based on sets of matches specified in \code{man}, using the join type
specified by argument \code{type} (see \link{join_types} for more details)
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using a dictionary
of manually-specified matches.
}
\examples{
data(ne_raw)
data(ne_ref)

# create df mapping sets of raw hierarchical values to codes within ref
ne_man <- data.frame(
  adm0 = NA_character_,
  adm1 = NA_character_,
  adm2 = "Bergen, N.J.",
  hcode = "211",
  stringsAsFactors = FALSE
)

# find manual matches
hmatch_manual(ne_raw, ne_ref, ne_man, code_col = "hcode", type = "inner")

}
