% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_parents.R
\name{hmatch_parents}
\alias{hmatch_parents}
\title{Hierarchical matching of parents based on sets of common offspring}
\usage{
hmatch_parents(
  x,
  raw,
  ref,
  level,
  mmin,
  pmin,
  pattern = NULL,
  pattern_ref = pattern,
  by = NULL,
  by_ref = by,
  type = "left",
  ref_prefix = "ref_",
  fuzzy = FALSE,
  max_dist = 1L,
  std_fn = string_std,
  ...
)
}
\arguments{
\item{x}{a \code{data.frame} representing one or more rows from \code{raw}, and
optionally including matching columns from \code{ref} to help narrow down the
set of possible offspring to try matching}

\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{level}{integer index of the hierarchical level to match at}

\item{mmin}{minimum number of matching offspring required for parents to be
considered a match}

\item{pmin}{minimum proportion of matching offspring required for parents to
be considered a match (i.e. the proportion of offspring in \code{raw} with match
in \code{ref})}

\item{pattern}{regex pattern to match the hierarchical columns in \code{raw}
(columns can be matched using either the \code{pattern} or \code{by} arguments, as
described in \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(defaults to \code{pattern})}

\item{by}{vector giving the names of the hierarchical columns in \code{raw}}

\item{by_ref}{vector giving the names of the hierarchical columns in \code{ref}
(defaults to \code{by})}

\item{type}{type of join ("inner" or "left") (defaults to "left")}

\item{ref_prefix}{Prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to \code{ref_})}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{Additional arguments passed to \code{std_fn()}}
}
\value{
a data frame obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, using the join type specified by argument \code{type} (see
\link{join_types} for more details)
}
\description{
Match a hierarchical column (e.g. region, province, or county) within a raw,
potentially messy dataset against a corresponding column within a reference
dataset, by searching for similar sets of 'offspring' (i.e. values at the
next hierarchical level).

For example, if the raw dataset uses admin1 level "NY" whereas the reference
dataset uses "New York", it would be difficult to automatically match these
values using only fuzzy-matching. However, we might nonetheless be able to
match "NY" to "New York" if they share a common and unique set of 'offspring'
(i.e. admin2 values) across both datasets (e.g "Kings", "Queens", "New York",
"Suffolk", "Bronx", etc.).
}
\examples{
# prepare example data
raw <- ne_ref
raw$adm1[raw$adm1 == "Ontario"] <- "ON"
raw$adm1[raw$adm1 == "New York"] <- "NY"
raw$adm1[raw$adm1 == "New Jersey"] <- "NJ"
raw$adm1[raw$adm1 == "Pennsylvania"] <- "PA"
x <- raw[raw$level == 1,]

hmatch_parents(
  x,
  raw,
  ne_ref,
  pattern = "adm",
  level = 2,
  mmin = 2,
  pmin = 0.5,
  type = "inner"
)

}
