% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_partial.R
\name{hmatch_partial}
\alias{hmatch_partial}
\title{Partial hierarchical matching}
\usage{
hmatch_partial(
  raw,
  ref,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  by = NULL,
  dict = NULL,
  type = "left",
  ref_prefix = "ref_",
  fuzzy = FALSE,
  max_dist = 1L,
  std_fn = string_std,
  ...
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}
(see also \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(see also \link{specifying_columns})}

\item{by}{named character vector whose elements are the names of the
hierarchical columns in \code{ref} and whose names are the names of the
corresponding columns in \code{raw} (see also \link{specifying_columns})}

\item{dict}{Optional dictionary for recoding values within the hierarchical
columns of \code{raw} (see \link{dictionary_recoding})}

\item{type}{type of join ("left", "inner", "inner_unique", "anti", or
"anti_unique"). Defaults to "left". See \link{join_types}.}

\item{ref_prefix}{Prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to \code{ref_})}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{Additional arguments passed to \code{std_fn()}}
}
\value{
a data frame obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, using the join type specified by argument \code{type} (see
\link{join_types} for more details)
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using partial
matching. "Partial" here means that one or more hierarchical levels within
the raw data may be missing (i.e. NA). More specifically, for a given row of
raw data, matches can potentially be made to a high-resolution level (e.g.
township) even if one or more lower-resolution levels (e.g. province) is
missing.
}
\examples{
data(ne_raw)
data(ne_ref)

hmatch_partial(ne_raw, ne_ref, pattern_raw = "adm")

# with dictionary-based recoding
ne_dict <- data.frame(value = "USA",
                      replacement = "United States",
                      variable = "adm0")

hmatch_partial(ne_raw, ne_ref, pattern_raw = "adm", dict = ne_dict)

}
