% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_partial.R
\name{hmatch_partial}
\alias{hmatch_partial}
\title{Partial hierarchical matching}
\usage{
hmatch_partial(
  raw,
  ref,
  pattern_raw = NULL,
  pattern_ref = pattern_raw,
  by = NULL,
  type = "left",
  std_fn = string_std,
  fuzzy = FALSE,
  max_dist = 1L
)
}
\arguments{
\item{raw}{\code{data.frame} containing hierarchical columns with raw, potentially
messy data}

\item{ref}{\code{data.frame} containing hierarchical columns with reference data}

\item{pattern_raw}{regex pattern to match the hierarchical columns in \code{raw}
(see also \link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(see also \link{specifying_columns})}

\item{by}{named character vector whose elements are the names of the
hierarchical columns in \code{ref} and whose names are the names of the
corresponding columns in \code{raw} (see also \link{specifying_columns})}

\item{type}{type of join ("inner" or "left") (defaults to "left")}

\item{std_fn}{Function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (defaults to
FALSE)}

\item{max_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use when
fuzzy-matching (defaults to \code{1L})}
}
\value{
A \code{data.frame} obtained by matching the hierarchical columns in \code{raw}
and \code{ref}. If \code{type == "inner"}, returns only the rows of \code{raw} with a
single match in \code{ref}. If \code{type == "left"}, returns all rows of \code{raw}. If
the hierarchical columns within \code{ref} have identical names to \code{raw}, the
returned reference columns will be renamed with prefix "bind_".
}
\description{
Match a data.frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using partial
matching. "Partial" here means that one or more hierarchical levels within
the raw data may be missing (i.e. NA). More specifically, for a given row of
raw data, matches can potentially be made to a high-resolution level (e.g.
township) even if one or more lower-resolution levels (e.g. province) is
missing.
}
\examples{
data(drc_raw)
data(drc_ref)

hmatch_partial(drc_raw, drc_ref, pattern_raw = "adm")

}
