% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatch_permute.R
\name{hmatch_permute}
\alias{hmatch_permute}
\title{Hierarchical matching with sequential column permutation to allow for values
entered at the wrong hierarchical level}
\usage{
hmatch_permute(
  raw,
  ref,
  pattern = NULL,
  pattern_ref = pattern,
  by = NULL,
  by_ref = by,
  type = "left",
  dict = NULL,
  ref_prefix = "ref_",
  fuzzy = FALSE,
  fuzzy_method = "osa",
  fuzzy_dist = 1L,
  std_fn = string_std,
  ...
)
}
\arguments{
\item{raw}{data frame containing hierarchical columns with raw data}

\item{ref}{data frame containing hierarchical columns with reference data}

\item{pattern}{regex pattern to match the hierarchical columns in \code{raw}
(columns can be matched using either \code{pattern} \emph{or} \code{by}, as described in
\link{specifying_columns})}

\item{pattern_ref}{regex pattern to match the hierarchical columns in \code{ref}
(defaults to \code{pattern})}

\item{by}{vector giving the names of the hierarchical columns in \code{raw}}

\item{by_ref}{vector giving the names of the hierarchical columns in \code{ref}
(defaults to \code{by})}

\item{type}{type of join ("left", "inner", "anti", "resolve_left",
"resolve_inner", or "resolve_anti"). Defaults to "left". See
\link{join_types}.}

\item{dict}{optional dictionary for recoding values within the hierarchical
columns of \code{raw} (see \link{dictionary_recoding})}

\item{ref_prefix}{prefix to add to hierarchical column names in \code{ref} if they
are otherwise identical to names in \code{raw}  (defaults to "ref_")}

\item{fuzzy}{logical indicating whether to use fuzzy-matching (based on the
\code{\link{stringdist}} package). Defaults to FALSE.}

\item{fuzzy_method}{if \code{fuzzy = TRUE}, the method to use for string distance
calculation (see \link[stringdist]{stringdist-metrics}). Defaults to "osa".}

\item{fuzzy_dist}{if \code{fuzzy = TRUE}, the maximum string distance to use to
classify matches (i.e. a string distance ≤ \code{fuzzy_dist} will be considered
matching). Defaults to \code{1L}.}

\item{std_fn}{function to standardize strings during matching. Defaults to
\code{\link{string_std}}. Set to \code{NULL} to omit standardization. See
also \link{string_standardization}.}

\item{...}{additional arguments passed to \code{std_fn()}}
}
\value{
a data frame obtained by matching the hierarchical columns in \code{raw}
and \code{ref}, using the join type specified by argument \code{type} (see
\link{join_types} for more details)
}
\description{
Match a data frame with raw, potentially messy hierarchical data (e.g.
province, county, township) against a reference dataset, using sequential
permutation of the hierarchical columns to allow for values entered at the
wrong hierarchical level.

The function calls \code{\link{hmatch}} on each possible permutation of the
hierarchical columns, and then combines the results. Rows of \code{raw} yielding
multiple matches to \code{ref} can optionally be resolved using a resolve-type
join (see section \strong{Resolve joins} below).
}
\section{Resolve joins}{

In \code{hmatch_permute}, if argument \code{type} corresponds to a resolve join, rows
of \code{raw} with multiple matches to \code{ref} are resolved to the highest
hierarchical level that is common among all matches (or no match if there is
a conflict at the very first level). E.g.

\code{raw}: \cr
\verb{1. | United States | <NA>     | New York |} \cr

Relevant rows from \code{ref}: \cr
\verb{1. | United States | New York | <NA>     |} \cr
\verb{2. | United States | New York | New York |} \cr

In a regular join with \code{hmatch_permute}, the single row from \code{raw} (above)
will match both of the depicted rows from \code{ref}. However, in a resolve join
the two matches will resolve to the first row from \code{ref}, because it reflects
the highest hierarchical level that is common to all matches.
}

\examples{
data(ne_raw)
data(ne_ref)

hmatch_permute(ne_raw, ne_ref, pattern = "^adm", type = "inner")

}
