% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ne_raw}
\alias{ne_raw}
\title{Raw dataset}
\format{
A data.frame with 12 rows and 4 variables:
\describe{
\item{id}{Integer id}
\item{adm0}{Name of administrative 0 level (country)}
\item{adm1}{Name of administrative 1 level (state/province)}
\item{adm2}{Name of administrative 2 level (county/census division)}
}
}
\usage{
ne_raw
}
\description{
Raw entries of select administrative districts from the northeastern portion
of North America.
}
\keyword{datasets}
