% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_std.R
\name{string_std}
\alias{string_std}
\title{String standardization prior to matching}
\usage{
string_std(x)
}
\arguments{
\item{x}{a string}
}
\value{
The standardized version of \code{x}
}
\description{
Standardizes strings prior to performing a match, using three
transformations:
\enumerate{
\item standardize case (\code{base::tolower})
\item replace sequences of space or punctuation characters with "_"
\item remove diacritics (\code{stringi::stri_trans_general})
}
}
\examples{
string_std("United STATES")
string_std("R\u00e9publique  d\u00e9mocratique du  Congo")

}
\seealso{
\link{string_standardization}
}
